feat: implement editable fields for event title and message with hover functionality
